# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "cutt.ovh Backend",
        "title": "cutt.ovh Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/v1",
    "paths": {
        "/redirect": {
            "post": {
                "description": "Adds new short url redirect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "Add redirect",
                "operationId": "postRedirect",
                "parameters": [
                    {
                        "description": "Set up url to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/redirect.postRedirectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/redirect.postRedirectResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/redirect.redirectErrorResponse"
                        }
                    }
                }
            }
        },
        "/redirect/{slug}": {
            "get": {
                "description": "Gets shortened url info by slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "Get redirect",
                "operationId": "getRedirectBySlug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/redirect.getRedirectBySlugResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/redirect.redirectErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/redirect.redirectErrorResponse"
                        }
                    }
                }
            }
        },
        "/redirect/{slug}/execute": {
            "get": {
                "description": "Executes redirect to original url",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "Get execute redirect",
                "operationId": "getRedirectExecute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": ""
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/redirect.redirectErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/redirect.redirectErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "redirect.getRedirectBySlugResponse": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "redirect.postRedirectRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "redirect.postRedirectResponse": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "redirect.redirectErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}